[#166][#183][#184] Add MIT liscence to new classes + refactor homecontroller to authentication
